Delete unnecessary code in secodnary input processing
Still had relic of code (preserved for now by commenting out) of using the pointer activation gesture with the secondary controller to activate the UVGR menu and then toggling the detection layers for the masks of the pointers.  At some point soon still need to implement the unique button combination for the UGVR that does not depend on the pointer being active or not.